pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "luthor-ai-systems-framework"
version = "0.1.0"
description = "Governance-first, deterministic AI frameworks for enterprise deployment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
  "httpx>=0.27"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
